+++ SaveShipItemsOnDeathMod/Patches/MainPatch.cs
@@ -1,14 +1,8 @@
-﻿using System;
-using System.Linq;
-using BepInEx;
-using GameNetcodeStuff;
+﻿using GameNetcodeStuff;
 using HarmonyLib;
-using Unity.Netcode;
-using UnityEngine;
 
 namespace SaveShipItemsOnDeathMod.Patches
 {
-    // TODO: Review/clean-up logs
     [HarmonyPatch(typeof(PlayerControllerB))]
     internal class MainPatch
     {
@@ -16,7 +10,7 @@ namespace SaveShipItemsOnDeathMod.Patches
         [HarmonyPostfix]
         public static void PostFillEndGameStatsHook(HUDManager __instance)
         {
-            ModLogger.Instance.LogInfo("Disabling allPlayersDead overlay");
+            ModLogger.Instance.LogDebug("Disabling allPlayersDead overlay");
             __instance.statsUIElements.allPlayersDeadOverlay.enabled = false;
         }
 
@@ -24,8 +18,8 @@ namespace SaveShipItemsOnDeathMod.Patches
         [HarmonyPostfix]
         public static void ShowSavedItemsNotificationOnPurpose()
         {
-            ModLogger.Instance.LogInfo("StartOfRound.AllPlayersHaveRevivedClientRpc patch");
-            ModLogger.Instance.LogInfo($"ShouldShowSavedItemsNotification? {ModVariables.Instance.ShouldShowSavedItemsNotification}");
+            ModLogger.Instance.LogDebug("StartOfRound.AllPlayersHaveRevivedClientRpc patch");
+            ModLogger.Instance.LogDebug($"ShouldShowSavedItemsNotification? {ModVariables.Instance.ShouldShowSavedItemsNotification}");
             
             if (ModVariables.Instance.ShouldShowSavedItemsNotification)
             {
@@ -47,14 +41,14 @@ namespace SaveShipItemsOnDeathMod.Patches
             
             if (TimeOfDay.Instance.daysUntilDeadline == 0)
             {
-                ModLogger.Instance.LogInfo("Ignore patch logic. Days until deadline = 0.");
+                ModLogger.Instance.LogDebug("Ignore patch logic. Days until deadline = 0.");
                 return;
             }
             
             if (StartOfRound.Instance.allPlayersDead)
             {
                 StartOfRound.Instance.allPlayersDead = false;
-                ModLogger.Instance.LogInfo($"Pre DespawnPropsAtEndOfRound, set allPlayersDead={StartOfRound.Instance.allPlayersDead}");
+                ModLogger.Instance.LogDebug($"Pre DespawnPropsAtEndOfRound, set allPlayersDead={StartOfRound.Instance.allPlayersDead}");
                 ModVariables.Instance.IsAllPlayersDeadOverride = true;
             }
         }
@@ -70,7 +64,8 @@ namespace SaveShipItemsOnDeathMod.Patches
             
             if (ModVariables.Instance.IsAllPlayersDeadOverride)
             {
-                var penaltyResult = PenaltyApplier.Apply();
+                var feePercent = SaveShipItemsOnDeathModNetworkManager.Instance.ScrapFee_Percent.Value;
+                var penaltyResult = PenaltyApplier.Apply(feePercent);
 
                 if (penaltyResult.IsError)
                 {
@@ -80,12 +75,8 @@ namespace SaveShipItemsOnDeathMod.Patches
 
                 __instance.allPlayersDead = true;
                 ModVariables.Instance.IsAllPlayersDeadOverride = false;
-                ModLogger.Instance.LogInfo($"Post DespawnPropsAtEndOfRound, set allPlayersDead={StartOfRound.Instance.allPlayersDead}");
+                ModLogger.Instance.LogDebug($"Post DespawnPropsAtEndOfRound, set allPlayersDead={StartOfRound.Instance.allPlayersDead}");
                 
-                // TODO: Implement here check, if server count not 0 but client,
-                // then take string[] of names of objects from server (should be also sent via RPC)
-                // and try to map it to names
-                // issue is that mod allows to connect after lobby closed, so player connected has items as InShip=false
                 if (penaltyResult.TotalItemsCount == 0)
                 {
                     return;
@@ -93,15 +84,18 @@ namespace SaveShipItemsOnDeathMod.Patches
 
                 var title = "KIRPICHYOV IND. MESSAGE";
                 var message = "Kirpichyov Ind. saved your items but have taken fees. " +
-                              "Scrap prices were cut in a half. " +
+                              $"Scrap prices were cut by {feePercent}%. " +
                               $"Total was {penaltyResult.TotalCostInitial}, now {penaltyResult.TotalCostCurrent}";
                 
                 HUDManager.Instance.AddTextToChatOnServer($"[Notification] {message}");
+
+                var updatedNetworkIdsString = string.Join(';', penaltyResult.UpdatedItemsNetworkIds);
                 
                 SaveShipItemsOnDeathModNetworkManager.Instance.ApplyItemsPenaltyClientRpc(
                     serverTotalCurrentCost: penaltyResult.TotalCostCurrent,
                     serverTotalItemsCount: penaltyResult.TotalItemsCount,
-                    serverTotalInitialCost: penaltyResult.TotalCostInitial);
+                    serverTotalInitialCost: penaltyResult.TotalCostInitial,
+                    serverNetworkObjectIds: updatedNetworkIdsString);
                 
                 SaveShipItemsOnDeathModNetworkManager.Instance.ShowItemsSavedNotificationOnReviveClientRpc(title, message);
             }