+++ ThunderstorePublish/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## v1.0.0
 - Initial public version
+
 ## v1.0.1
 - Reworked mod to use Server-Clients relationship. Now fees will be applied to items on the server side first and then sync across the clients
 - Removed items saved message in the stats screen on death that was visible only to host and added a tip notification on revive instead
+
+# v.1.0.2
+- Added ability to configure the scrap penalty fee percentage
+- Fixed visual scrap cost sync issue for clients when server already has items that were placed on ship before clients connected